pyproject.toml
[project]
name = "demo-finance-engine"
version = "0.1.0"
description = "Personal Finance Data Platform with Apache Spark & Delta Lake"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Noah", email = "noah@example.com"}
]
keywords = ["spark", "delta-lake", "data-engineering", "etl", "medallion-architecture"]

dependencies = [
    "pyspark==3.5.0",
    "delta-spark==3.0.0",
    "faker==22.0.0",
    "pandas>=2.1.4",
    "pyarrow>=14.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
]
quality = [
    "great-expectations>=0.18.8",
]
viz = [
    "jupyter>=1.0.0",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jobs", "utils", "data_generator", "schemas", "analytics"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=. --cov-report=html --cov-report=term"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

